Google button is good now and company summary done

+++ client/src/App.js
@@ -40,7 +40,8 @@ const App = () => {
         })
     } catch (err) {
         console.log("ERROR: ", err)
-        if(err.response.status === 401) {
+        console.log(err.response !== undefined)
+        if(err.response !== undefined && err.response.status === 401) {
           localStorage.clear();
 	        userCtx.setDefault();
         } else {
@@ -55,7 +56,6 @@ const App = () => {
     }
 
     if(localStorage.getItem("email") !== null) {
-      //localStorage.setItem("access_token", "I bet bet bet")
       console.log("ACCESS TOKEN:", localStorage.getItem("access_token"))
       userInfo.balance = accountFromServer.data.balance
       // console.log("userInfo in APP.js, from persistent login: ", userInfo)
